Refactor to source from aurae monorepo
