+++ pkg/runtime/runtime_client.go
@@ -2,7 +2,7 @@ package runtime
 
 import (
 	"context"
-	"github.com/aurae-runtime/client-go/pkg/stdlib/v0/runtime"
+	"github.com/aurae-runtime/client-go/pkg/api/v0/runtime"
 	"google.golang.org/grpc"
 )
