+++ pkg/schedule/schedule_client.go
@@ -2,7 +2,7 @@ package schedule
 
 import (
 	"context"
-	"github.com/aurae-runtime/client-go/pkg/stdlib/v0/schedule"
+	"github.com/aurae-runtime/client-go/pkg/api/v0/schedule"
 	"google.golang.org/grpc"
 )
